Refactor game code and enhance gameplay
This comprehensive commit brings to baseline several changes aimed at refining the codebase and enhancing our game's gameplay. The prueba.py file was deleted as it was outdated and unnecessarily cluttering the codebase, and the functionality was merged and better organized in other files.

Changes made:
- In jugador.py, the player has now vertical movements, number of lives increased, modified the shooting mechanism to limit the number of bullets based on a new attribute, created an Enemy class for enemy sprites, handling Enemy movements and updated collision detections.
- In menu.py, the text and placement of text on the menu and credits screens have been adjusted for better visibility and understandability.
- In jugabilidad.py, introduced enemy generation logic, added game over condition, updated events handling and render method. Player's projectiles and collisions are now updated here.
- In config.py, added paths of enemy images to support multiple appearances for enemies.
- Introduced menu and HUD classes in menu.py to handle game interface cleanly.

These changes collectively enhance the fluidity and variability of the game, increase the readability of the code, and make it easier to track and handle game elements state in jugabilidad. It adds to the fun and attractiveness of the game and improves the user experience.